Updated the format of relationships
the relations now accept either a string or an object. the object holds
the specification of the relationship in the format:

hasMany( { resource: "friends",
           type: "persons" } )

If the type of the relationship is the same as the name of the
relationship, type can be omitted in the specification object.

Passing

hasMany( "persons" )

is the same as

hasMany( { resource: "persons",
           type: "persons" } )